fix(types): Add CodecMap type alias (#56)
