+++ src/index.ts
@@ -8,7 +8,13 @@ export {
   CodecsType,
   DecodeQuery,
 } from "./lib/Codecs";
-export { CodecsToRecord, InferQueryParams, PathLike, RouteParams } from "./lib/commons.types";
+export {
+  CodecMap,
+  CodecsToRecord,
+  InferQueryParams,
+  PathLike,
+  RouteParams,
+} from "./lib/helpers/common";
 export { CodecDecodeError } from "./lib/errors/CodecDecodeError";
 export { CodecEncodeError } from "./lib/errors/CodecEncodeError";
 export { UrlParserError } from "./lib/errors/UrlParserError";
