+++ src/lib/react/hookMakers.ts
@@ -1,20 +1,18 @@
 import { Dispatch, SetStateAction, useEffect, useMemo, useState } from "react";
 
-import { Codec } from "../Codecs";
-import { CodecsToRecord, PathLike, RouteParams } from "../commons.types";
-import { safeKeys } from "../helpers/commons";
+import { CodecMap, CodecsToRecord, PathLike, RouteParams, safeKeys } from "../helpers/common";
 import { Routeway } from "../Routeways";
 
 type RouteParamsHook = <
-  V extends Record<string, Codec<unknown>>,
-  Q extends Record<string, Codec<unknown>>,
+  V extends CodecMap,
+  Q extends CodecMap,
 >(route: Routeway<PathLike, V, Q>) => {
   pathVars: CodecsToRecord<V>;
   queryParams: Partial<CodecsToRecord<Q>>;
   setQueryParams: Dispatch<SetStateAction<Partial<CodecsToRecord<Q>>>>;
 };
 
-type PathVarsHook = <V extends Record<string, Codec<unknown>>>(route: Routeway<PathLike, V>) => CodecsToRecord<V>;
+type PathVarsHook = <V extends CodecMap>(route: Routeway<PathLike, V>) => CodecsToRecord<V>;
 
 interface QueryParamHook {
   /**
@@ -26,8 +24,8 @@ interface QueryParamHook {
    * @param key the key of the specific query param
    * @returns a React state and dispatcher tuple of the query param
    */
-  <Q extends Record<string, Codec<unknown>>, K extends keyof Q>(
-    route: Routeway<PathLike, Record<string, Codec<unknown>>, Q>,
+  <Q extends CodecMap, K extends keyof Q>(
+    route: Routeway<PathLike, CodecMap, Q>,
     key: K,
   ): [
     Partial<Q>[K],
@@ -43,8 +41,8 @@ interface QueryParamHook {
    * @param fallback a value to fall back if the query param is `undefined`
    * @returns a React state and dispatcher tuple of the query param
    */
-  <Q extends Record<string, Codec<unknown>>, K extends keyof Q>(
-    route: Routeway<PathLike, Record<string, Codec<unknown>>, Q>,
+  <Q extends CodecMap, K extends keyof Q>(
+    route: Routeway<PathLike, CodecMap, Q>,
     key: K,
     fallback: NonNullable<Partial<Q>[K]>,
   ): [
@@ -53,8 +51,8 @@ interface QueryParamHook {
   ];
 }
 
-type AllQueryParamsHook = <Q extends Record<string, Codec<unknown>>>(
-  route: Routeway<PathLike, Record<string, Codec<unknown>>, Q>
+type AllQueryParamsHook = <Q extends CodecMap>(
+  route: Routeway<PathLike, CodecMap, Q>
 ) => {
   queryParams: Partial<CodecsToRecord<Q>>;
   setQueryParams: Dispatch<SetStateAction<Partial<CodecsToRecord<Q>>>>;
@@ -73,8 +71,8 @@ interface LocationLike {
 }
 
 type NavigateMethods<
-  V extends Record<string, Codec<unknown>>,
-  Q extends Record<string, Codec<unknown>>,
+  V extends CodecMap,
+  Q extends CodecMap,
 > = keyof V extends never
       ? {
           navigate(params?: RouteParams<V, Q>): void;
