+++ test/unit/lib/Commons.types.test.ts
@@ -1,7 +1,7 @@
 import { expect } from "@stackbuilders/assertive-ts";
 
 import { Codecs } from "../../../src/lib/Codecs";
-import { CodecsToRecord, InferQueryParams, PathLike, RouteParams } from "../../../src/lib/commons.types";
+import { CodecsToRecord, InferQueryParams, PathLike, RouteParams } from "../../../src/lib/helpers/common";
 import { TestRoutes } from "../../TestRoutes";
 
 describe("[Unit] Commons.types.test.ts", () => {
